Move commandline interface to typer
This commit moves the kiwi commandline and plugin interface
to typer and drops the use of docopt. Typer is based on python
type hints and allows for a more modern way to define Cli
interfaces and auto completion, even for plugins. Also docopt
is split into an old and a next generation variant and allows
us to get rid of spec file hacks as well as documentation
rendering hacks.