Seed filesystem UUIDs with label info
For reproducible builds the calculation of the filesystem UUID
should be persistent with each rebuild of the image. To achieve
this the UUID is calculated using the label name as seed. Only
in case the filesystem has no label assigned the UUID stays
random and changes with each rebuild. In addition to this change
a new label attribute has been added to the <partition> section.
This Fixes #2761