Add handling for stale issues and PRs
As discussed in the community meeting stale issues and PRs
will be warned and stale labeled after 60 days. In the
community meeting we will go over all stale PRs and close
them on agreement. We will also look on stale issues and
decide if closing them is useful or not. There will never
be an auto-close.