Drop root only resize restriction
The current kiwi-resize code was restricted to the root
partition for historical reasons. As in a partition table
only the last partition can resize, this should be the only
limitation for the resize code to perform its job. In
connection with the rd.kiwi.install.retain_last feature
it is also very likely that the last partition is not
the root partition but it should be properly restored by
the resize code after deployment