Add rd.kiwi.install.retain_past deployment option
As a less strict method to retain_last, now also the retain_past
deployment option exists. In the retain_past mode the partition
to preserve must not be part of the initial image. It can retain
the content of any last partition if its start address is past
the OS image. However, a gap in the partition table between the
OS end and the start of the retained partition can occur for
which no resize method exists yet. In addition there is no
sanity checking when there are multiple partitions, not just
only one last after the OS image. As such this feature must
be used with caution.